Adds logging handlers

+++ climetar/__init__.py
@@ -1,12 +1,20 @@
+# Imports logging posibilities
+from .logger import getLogger
+logger = getLogger(__name__)
 
 #Prevents the 'Bad Key "text.kerning_factor" on line 4' warning when working with matplotlib
-with open('/opt/conda/lib/python3.7/site-packages/matplotlib/mpl-data/stylelib/_classic_test_patch.mplstyle','r+') as fh:
-    lines = fh.readlines()
-    lines = ['#'+line for line in lines if line.startswith("text.kerning_factor")]
-    fh.seek(0)
-    fh.writelines(lines)
-    fh.truncate()
+import os
+metarfile = '/opt/conda/lib/python3.7/site-packages/matplotlib/mpl-data/stylelib/_classic_test_patch.mplstyle'
+if os.path.isfile(metarfile):
+    with open(metarfile,'r+') as fh:
+        lines = fh.readlines()
+        lines = ['#'+line for line in lines if line.startswith("text.kerning_factor")]
+        fh.seek(0)
+        fh.writelines(lines)
+        fh.truncate()
 
+
+# Imports climetar modules
 from .quantities import *
 from .metar import *
 
@@ -18,4 +26,3 @@ from .metarfile import MetarFiles, analyse_files, format_period
 from .planetary import Astro
 
 from .metarplot import MetarPlotter
-
